Gave Scripting Classes their own files
