+++ GameEngine/src/Engine/Scripting/ScriptEngine.h
@@ -15,83 +15,13 @@ extern "C"
 	typedef struct _MonoString MonoString;
 }
 
-typedef void(*OnCreate) (MonoObject* obj, MonoObject** exp);
-typedef void(*OnDestroy) (MonoObject* obj, MonoObject** exp);
-typedef void(*OnUpdate) (MonoObject* obj, MonoObject* timestep, MonoObject** exp);
-
 // Created with help from this guide (Mono Embedding for Game Engines): https://peter1745.github.io/introduction.html
 
 namespace Engine
 {
-
+	// forward declaration
+	class ScriptClass;
 	class ScriptInstance;
-	class ScriptField
-	{
-	public:
-		ScriptField() = default;
-		ScriptField(MonoClassField* monoField);
-		~ScriptField() = default;
-
-		MonoClassField* GetMonoField() { return m_MonoField; }
-		const std::string& GetTypeName() { return m_TypeName; }
-		void GetValue(Ref<ScriptInstance> instance, void* value);
-		void SetValue(Ref<ScriptInstance> instance, void* value);
-
-		bool IsPublic();
-
-	private:
-		MonoClassField* m_MonoField = nullptr;
-		uint8_t m_Access;
-		std::string m_TypeName;
-	};
-
-	class ScriptClass
-	{
-	public:
-		ScriptClass(const std::string& classNamespace, const std::string& className);
-		~ScriptClass() = default;
-
-		MonoObject* Instantiate();
-
-		MonoClass* GetMonoClass() { return m_MonoClass; }
-
-		MonoMethod* GetMethod(const std::string& name, int parameterCount);
-		MonoObject* InvokeMethod(MonoObject* instance, MonoMethod* method, void** params = nullptr);
-
-		Ref<ScriptField> GetScriptField(const std::string& fieldName) { return m_ScriptFields[fieldName]; }
-		std::unordered_map<std::string, Ref<ScriptField>> GetScriptFields() { return m_ScriptFields; }
-
-	private:
-		std::string m_ClassNamespace;
-		std::string m_ClassName;
-
-		MonoClass* m_MonoClass = nullptr;
-		std::unordered_map<std::string, Ref<ScriptField>> m_ScriptFields;
-	};
-
-	class ScriptInstance
-	{
-	public:
-		ScriptInstance() = default;
-		ScriptInstance(Ref<ScriptClass> scriptClass, const UUID& entityID);
-		~ScriptInstance() = default;
-
-		void InvokeOnCreate();
-		void InvokeOnDestroy();
-		void InvokeOnUpdate(Timestep ts);
-
-		MonoObject* GetMonoObject() { return m_Instance; }
-
-	private:
-		Ref<ScriptClass> m_ScriptClass;
-
-		MonoObject* m_Instance = nullptr;
-
-		MonoMethod* m_Constructor = nullptr;
-		OnCreate OnCreateThunk = nullptr;
-		OnDestroy OnDestroyThunk = nullptr;
-		OnUpdate OnUpdateThunk = nullptr;
-	};
 
 	class ScriptEngine
 	{
@@ -121,12 +51,25 @@ namespace Engine
 		static MonoDomain* GetRootDomain();
 		static MonoDomain* GetAppDomain();
 		static MonoAssembly* GetCoreAssembly();
+		static MonoAssembly* GetAppAssembly();
+		static MonoClass* GetEntityClass();
+		static MonoClass* GetTimestepClass();
 
 		static MonoClass* GetClassInAssembly(MonoAssembly* assembly, const char* namespaceName, const char* className);
 		static uint8_t GetFieldAccessibility(MonoClassField* field);
 		static uint8_t GetPropertyAccessbility(MonoProperty* property);
 
+		static void HandleMonoException(MonoObject* ptrExObject);
 		static std::string MonoStringToUTF8(MonoString* monoString);
+
+		enum class Accessibility : uint8_t
+		{
+			None = 0,
+			Private = (1 << 0),
+			Internal = (1 << 1),
+			Protected = (1 << 2),
+			Public = (1 << 3)
+		};
 
 	private:
 		static void InitMono();
